Use claude to automatically organize improvement PRs
Claude code fetched all the improvement PR content from github and used
that to do the categorization. It used my previous WIP commit as a
starting point. I ran it from inside the Slicer codebase so that it has
the context of the CLAUDE.md that it generated fro slicer. I also had it
output summaries into slicer-5.10-pr-summaries.md since it was fetching
the PRs anyway.